+++ test/test_helper.exs
@@ -32,6 +32,7 @@ ExUnit.start(
     # :pix_infraction,
     # :pix_infraction_log,
     # :pix_domain,
+    # :static_brcode,
     # :issuing_balance,
     # :issuing_product,
     # :issuing_card,
@@ -70,9 +71,9 @@ Code.require_file("./test/utils/pix_fraud.exs")
 Code.require_file("./test/utils/pix_dispute.exs")
 Code.require_file("./test/utils/pix_pull_request.exs")
 Code.require_file("./test/utils/random.exs")
+Code.require_file("./test/utils/static_brcode.exs")
 Code.require_file("./test/utils/ledger.exs")
 Code.require_file("./test/utils/ledger_transaction.exs")
 Code.require_file("./test/utils/issuing_holder.exs")
 Code.require_file("./test/utils/issuing_withdrawal.exs")
 Code.require_file("./test/utils/issuing_invoice.exs")
-Code.require_file("./test/utils/static_brcode.exs")