+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 ## [Unreleased]
 ### Added
+- StaticBrcode resource
 - PixPullRequest resource
 - PixPullRequest.Log sub-resource
 - Ledger resource